Adjust enemysnake base on the refactored code
